move package and fix error
